# Add gRPC client-side integration with golang.org/x/net/trace.