This will make client (outgoing) traces appear on /debug/requests.
There's some information lacking, but it's enough for now.