@iamqizhao 2015-06-05
Is it possible to add a end2end test for this?