@dsymonds 2015-06-08
Not just yet. I'm running benchmarks now that I've got those fixed.