@dsymonds 2015-06-08
Here's a benchmark comparison between `enableTracing = false` and `enableTracing = true`:

```
benchmark                     old ns/op     new ns/op     delta
BenchmarkClientStreamc1       78052         77161         -1.14%
BenchmarkClientStreamc8       52947         53064         +0.22%
BenchmarkClientStreamc64      51948         51674         -0.53%
BenchmarkClientStreamc512     57291         56697         -1.04%
BenchmarkClientUnaryc1        143743        150490        +4.69%
BenchmarkClientUnaryc8        113032        118492        +4.83%
BenchmarkClientUnaryc64       109480        114804        +4.86%
BenchmarkClientUnaryc512      101987        107993        +5.89%
```

I suspect the ~zero impact on streaming is because everything isn't hooked up, but it shows that the impact on this change is ~4-5%. That's relatively small for such a big feature, and is an even smaller impact on a program that isn't simply benchmarking the gRPC code.

I am happy to keep the unexported bool here to make it easy to benchmark this in the future, but I'd prefer to not export it right now. The API for this package is big enough, and it'd be nice to not grow the API unless there's clear impact that requires it. It's easy to come back later and export it if someone is actually having problems due to tracing.

PTAL.