@dsymonds 2015-06-08
Here's a comparison with 1 KB payloads. Note that I had to run with `-benchtime=10s` to get something even reasonably stable, though note that there's still a lot of noise. We're looking at small numbers here (~120-170 microseconds per op), and that's tiny compared to actual network speeds. You can even see that this makes the 1-concurrent version appear to get *faster* with tracing turned on. This is exercising the Go runtime as much as gRPC itself.

```
benchmark                     old ns/op     new ns/op     delta
BenchmarkClientUnaryc1        175516        170908        -2.63%
BenchmarkClientUnaryc8        125572        132750        +5.72%
BenchmarkClientUnaryc64       135132        143311        +6.05%
BenchmarkClientUnaryc512      158194        167593        +5.94%
```

I can try benchmarking across machines later today. I suspect the delta will appear very close to zero under even vaguely realistic situations.