@iamqizhao 2015-06-08
Hi David,

Thanks for doing this. For rpc frame benchmark (for both google internal
and grpc), we target micro-second optimization (for C++, sometimes we do
sub-micro optimization). One reason is that some usage is on the same host
via some IPC or even in process. For those case, network time is close to 0
and the efficiency of grpc runtime is critical. So persistent ~5% increase
is a big deal and we should take some action. Please have a switch with
default off. @yangzhouhan is the summer intern here working on grpc-go
performance. She can take a close look after your submission.

On Mon, Jun 8, 2015 at 12:03 PM, David Symonds <notifications@github.com>
wrote:

> Here's a comparison with 1 KB payloads. Note that I had to run with
> -benchtime=10s to get something even reasonably stable, though note that
> there's still a lot of noise. We're looking at small numbers here (~120-170
> microseconds per op), and that's tiny compared to actual network speeds.
> You can even see that this makes the 1-concurrent version appear to get
> *faster* with tracing turned on. This is exercising the Go runtime as
> much as gRPC itself.
>
> benchmark                     old ns/op     new ns/op     delta
> BenchmarkClientUnaryc1        175516        170908        -2.63%
> BenchmarkClientUnaryc8        125572        132750        +5.72%
> BenchmarkClientUnaryc64       135132        143311        +6.05%
> BenchmarkClientUnaryc512      158194        167593        +5.94%
>
> I can try benchmarking across machines later today. I suspect the delta
> will appear very close to zero under even vaguely realistic situations.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/grpc/grpc-go/pull/210#issuecomment-110108713>.
>