@dsymonds 2015-06-09
I've exported `EnableTracing`, defaulting to `true`. I think that's the right compromise. Almost all users of grpc-go will find ~7µs to be an utterly trivial amount of time for a big benefit, and the rare bit of code that both only does loopback networking and also almost no other work can set it to `false` easily enough.