@iamqizhao 2015-06-09
On Mon, Jun 8, 2015 at 6:14 PM, David Symonds <notifications@github.com>
wrote:

> I've exported EnableTracing, defaulting to true. I think that's the right
> compromise. Almost all users of grpc-go will find ~7µs to be an utterly
> trivial amount of time for a big benefit, and the rare bit of code that
> both only does loopback networking and also almost no other work can set it
> to false easily enough.
>
okay ... I do think it should be true eventually but probably not at this
time. Anyways, sgtm.

ready for review?

> —
> Reply to this email directly or view it on GitHub
> <https://github.com/grpc/grpc-go/pull/210#issuecomment-110188116>.
>