@dsymonds 2015-06-09
If it's eventually going to default to true, we should do that now, so it'll show up in benchmarks and profiles so we stay aware of it and don't let it get out of control.

Yes, this is ready for review.