@peter-edge 2015-07-17
@dsymonds this is probably a better PR to comment on. Is there any other documentation on this? I cannot seem to get tracing on my end.