@dsymonds 2015-07-17
The golang.org/x/net/trace package docs say to look at /debug/requests on a running program. Note that only client-side tracing has been implemented so far.