@peter-edge 2015-07-17
Have you successfully done this? What is it doing? Because I tried this and got an error. Do you mean that on whatever process a gRPC client is running, an http server will also implicitly be run, because this seems bad?

I've been using grpc-go for a while and I really have no idea how to use this feature, I really think some documentation/examples are needed.